TIP-221: rename AttributeManager to AttributeOptionsSorter
