+++ src/Pim/Bundle/CatalogBundle/Resources/config/managers.yml
@@ -1,5 +1,5 @@
 parameters:
-    pim_catalog.manager.attribute.class:           Pim\Bundle\CatalogBundle\Manager\AttributeManager
+    pim_catalog.manager.attribute_options_sorter.class: Pim\Bundle\CatalogBundle\Manager\AttributeOptionsSorter
     pim_catalog.manager.attribute_option.class:    Pim\Bundle\CatalogBundle\Manager\AttributeOptionManager
     pim_catalog.manager.attribute_group.class:     Pim\Bundle\CatalogBundle\Manager\AttributeGroupManager
     pim_catalog.manager.category.class:            Pim\Bundle\CatalogBundle\Manager\CategoryManager
@@ -36,14 +36,10 @@ services:
             - '@pim_catalog.factory.category'
             - %pim_catalog.entity.category.class%
 
-    pim_catalog.manager.attribute:
-        class: %pim_catalog.manager.attribute.class%
+    pim_catalog.manager.attribute_options_sorter:
+        class: %pim_catalog.manager.attribute_options_sorter.class%
         arguments:
-            - %pim_catalog.entity.attribute.class%
-            - '@pim_catalog.registry.attribute_type'
             - '@pim_catalog.saver.attribute_option'
-            - '@pim_catalog.repository.attribute'
-            - '@pim_catalog.factory.attribute'
 
     pim_catalog.manager.attribute_option:
         class: %pim_catalog.manager.attribute_option.class%
