+++ src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
@@ -151,10 +151,11 @@ services:
             - '@doctrine.orm.entity_manager'
             - '@form.factory'
             - '@fos_rest.view_handler'
-            - '@pim_catalog.manager.attribute'
+            - '@pim_catalog.manager.attribute_options_sorter'
             - '@pim_catalog.manager.attribute_option'
             - '@pim_catalog.saver.attribute_option'
             - '@pim_catalog.remover.attribute_option'
+            - '@pim_catalog.repository.attribute'
 
     pim_enrich.controller.product:
         scope: request
